+++ Effects/Package.swift
@@ -11,9 +11,6 @@ let package = Package(
         .library(
             name: "Effects",
             targets: ["Effects"]),
-        .library(
-            name: "LocalService",
-            targets: ["LocalService"]),
     ],
     dependencies: [
         .package(url: "https://github.com/pointfreeco/swift-composable-architecture", branch: "main"),
@@ -24,12 +21,6 @@ let package = Package(
             name: "Effects",
             dependencies: [
                 .product(name: "ComposableArchitecture", package: "swift-composable-architecture")
-            ]),
-        .target(
-            name: "LocalService",
-            dependencies: [
-                .product(name: "ComposableArchitecture", package: "swift-composable-architecture"),
-                "CombineRealm",
             ]),
     ]
 )
