--- Effects/Sources/LocalService/Common/BaseRealm.swift
@@ -1,7 +1,0 @@
-public class BaseService  {
-    unowned let realm : ServiceRealmType
-    
-    init(realm : ServiceRealmType){
-        self.realm = realm
-    }
-}
