--- Effects/Sources/LocalService/Common/RealmProvider.swift
@@ -1,9 +1,0 @@
-public protocol ServiceRealmType : AnyObject{
-    var realms : Realms {get}
-    
-}
-public final class ServiceRealm : ServiceRealmType{
-    public lazy var realms: Realms = RealmService(realm: self)
-    
-    public init(){}
-}
