+++ gitignore/Sources/Tabbar/TabbarFeacture.swift
@@ -7,7 +7,6 @@
 
 import ComposableArchitecture
 import Effects
-import LocalService
 
 struct TabBarState : Equatable{
     var selectedTab : String = "Home"
@@ -42,10 +41,7 @@ let tabBarReducer = Reducer<TabBarState, TabBarAction, tabBarEnvironmnet>.combin
         state:\.listState,
         action: /TabBarAction.listAction,
         environment: {_ in
-                .init(
-                    mainQueue: {.main},
-                    localService: { ServiceRealm()}
-                )
+                .init(mainQueue: {.main})
         }
     ),
     issueReducer.pullback(
