+++ gitignore/Sources/UI/List/ListFeacture.swift
@@ -1,5 +1,4 @@
 import ComposableArchitecture
-import LocalService
 
 struct ListState: Equatable{
     var listcount : Int = .init()
@@ -13,13 +12,10 @@ enum ListAction{
 struct ListEnvironmnet{
     
     var mainQueue: () -> AnySchedulerOf<DispatchQueue>
-    var localService : () -> ServiceRealmType
     public init(
-        mainQueue : @escaping() -> AnySchedulerOf<DispatchQueue>,
-        localService :  @escaping() -> ServiceRealmType
+        mainQueue : @escaping() -> AnySchedulerOf<DispatchQueue>
     ) {
         self.mainQueue = mainQueue
-        self.localService = localService
     }
 }
 
@@ -30,7 +26,6 @@ let listReducer = Reducer<
 >{ state ,action, environment in
     switch action{
     case .onAppear:
-        environment.localService()
         return .none
     }
