+++ gitignore/Sources/UI/List/ListView.swift
@@ -48,3 +48,16 @@ struct Card {
     let subtitle: String
     let tag : [String]
 }
+
+class ListModelex{
+    var title : String
+    var time = Date()
+    var tag : [String]
+    var bookMark : Bool
+    
+    init(title: String, tag : [String], bookMark : Bool){
+        self.title = title
+        self.tag = tag
+        self.bookMark = bookMark
+    }
+}
